requirements.txt
streamlit 
faiss-cpu 
sentence-transformers 
google-generativeai 
python-docx 
PyPDF2 
python-dotenv 
